- added day 13 - removed old project data
