+++ day12/build.gradle
@@ -1,3 +1,3 @@
 dependencies {
-    implementation("com.squareup.moshi:moshi-kotlin:1.8.0")
+    implementation "com.squareup.moshi:moshi-kotlin:1.8.0"
 }
